refactor(Timeline): extract styled components into own module
