+++ next/src/pages/Start.tsx
@@ -9,11 +9,11 @@ import EnrolmentView from "../components/start/enrolment/EnrolmentView";
 import ConnectView from "../components/start/connect/ConnectView";
 import EventsView from "../components/start/events/EventsView";
 import useTimelineScroll from "../hooks/TimelineScroll";
-import Timeline from "../components/start/Timeline";
+import Timeline from "../components/start/Timeline/Timeline";
 
 const MainContainer = styled.div`
   padding: 3vw 3vw;
-  font-family: 'Raleway';
+  font-family: "Raleway";
   font-weight: 450;
   font-size: 15px;
   flex: 1;
@@ -22,7 +22,7 @@ const MainContainer = styled.div`
     padding: 12vw 10vw;
     text-align: center;
   }
-`
+`;
 
 const PageContainer = styled.div`
   max-width: 100vw;
@@ -36,7 +36,7 @@ const views: Record<string, ReactNode> = {
   Connect: <ConnectView />,
   Advice: <AdviceView />,
   Enrolment: <EnrolmentView />,
-  Events: <EventsView />
+  Events: <EventsView />,
 };
 
 export default function Start() {
@@ -48,16 +48,16 @@ export default function Start() {
   const [scrolling, handleScroll, focusedView, setFocusedView] = useTimelineScroll(
     Object.keys(views).length,
     1000,
-    () => scrolling.current
+    () => scrolling.current,
   );
 
   return (
     <PageContainer>
       <Navbar open={navbarOpen} setNavbarOpen={handleToggle} variant={NavbarType.MINIPAGE} />
-      <MainContainer onWheel={(e) => handleScroll(e.deltaY)}>
+      <MainContainer onWheel={e => handleScroll(e.deltaY)}>
         {Object.values(views)[focusedView]}
       </MainContainer>
       <Timeline focusedView={focusedView} viewNames={Object.keys(views)} />
     </PageContainer>
-  )
+  );
 }
